Make builder typesafe by default and deprecate type-unsafe interfaces
This concludes the conversion to typesafe builders, by introducing two
new factory methods which can be used with top-level containers,
resulting in a type-initialized container, on which child() and
augmentation() methods work as expected.

Since using this instantiation guides the developer and provides useful
compile-time feedback on what constitutes a valid InstanceIdentifier,
deprecate the old, unsafe interface.

Signed-off-by: Robert Varga <rovarga@cisco.com>