Bump model versions to prevent clash with nexus
It turns out the current -SNAPSHOT version were release on 9/30 into
nexus, but their versions were not bumpted. The released version does
not carry the latest generation rules, most notably DTO range checks and
it is tied to a historic (0.1.1) version of model parent.

This means we need to fix the historic mistake and do a fresh release.

Change-Id: Ia15862d880833bfc49cb42dd298a45c9b252fb32
Signed-off-by: Robert Varga <rovarga@cisco.com>