Update patch set 1
Patch Set 1: Code-Review-1

This change is a bit more complicated that this. Boolean leaves in YANG are really tri-state: true, false and not-present. This patch in its current form breaks that 'not present' case.

This user API simplification can only be made if the leaf contains a 'default' directive, in which case the non-presence of an attribute should be turned either to 'true' or to 'false' -- based on that directive. I would argue that implementing it would be better in the builder default constructor and setter, which would make sure that the correct value is always set -- thus the getter would remain simple.

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1