Bug 157 fix - name conflict when YANG list has element with >key< name
According to RFC 6020 name of concrete container, leaf, leaf-list, list,
grouping, choice ... has to start with letter which is followed with
letters, digits, '_', '-', or '.' character.

Before this commit there was generated for key of a list getter method
'getKey'. 'key' word was used while getter method was constructed.
In this commit 'key$' word is used. Dollar is character which can't be used in
YANG but can be used in JAVA. Therefore it can't happen that list
subelement identifier will be also 'key$'.

As was specified in bugzilla - interface (from YANG list) and implementing
class are generated with 'key()' getter method and with member field '_key$'.
Builder for implementing class is generated with member field
'_key$' and with 'key()' and 'Builder key(parameter)' getter and setter methods.

Name conflict for methods 'key()' is solved
because all methods generated from subelements of YANG list starts with
'get', 'set' or 'is'.

Change-Id: I160a759c16f24c7e59d7f5de66a4d9b8a39ae7cd
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>