Use a random port in test server
This patch updates the test server to bind to a random port number and expose
this port number via a Future can be fetched by the test client so that it
knows to which port to connect.

Version 2:
Modified the code to use the guava SettableFuture as opposed to a custom built
future, as well as cleaned up a few other things, including makding sure that
the message was received from the server before considering the test passed.
was seeing different behavior on different platforms without this (i.e. some
times the message was display, sometimes not). Also, failed the test case wh
exceptions were caught.

Change-Id: Ia11c374bfbf206a6214d2afe3ba328b1231d6fad
Signed-off-by: David K. Bainbridge <dbainbri@ciena.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>