Introduce the concept of ProductAwareBuilder
A ProductAwareBuilder is an builder which can act as a proxy
for its product in collections (obviously provided it is not
mutated while its used this way).

This functionality is useful for implemeting caches, where
the toInstance() call can be skipped as long as there is an
existing equal object.

Change-Id: I24c98dcb41491ba6265b373c34667e2a0ddfed86
Signed-off-by: Robert Varga <rovarga@cisco.com>