BUG-648: Fixup hashCode/equals
The baseline hashCode/equals implementation relied on the copied maps
for correctness. Iterable<?> in and of itself does not really have a
equals contract, so let's reimplement it and make it very explicit.

This patch is required for getting rid of the copied maps without
breaking users. Note that this uncovers an issue where attributes were
not taken into account by equals() -- disable that test for now.

Change-Id: I2fcd2bc8775b60cf89bf03a5679f97a266cf1c1e
Signed-off-by: Robert Varga <rovarga@cisco.com>