Bug 595 - Failed to maps instance-identifier (XML node) into Java InstanceIdentifier
The case for instance identifier was added to
XmlDocumentUtils.toSimpleNodeWithType() method. If according to schema XML
element is InstanceIdentifier then it is attempted to deserialize as
InstanceIdentifier via specific codec - InstanceIdentifierForXmlCodec.
Deserialize method acceot as input parameter Element and Schema context.
Serialize method accept as input parameter InstanceIdentifier and Element.

Change-Id: I27b91e98479dfb4b0134294b5f3e608a8bf8ad4d
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>