Update patch set 2
Patch Set 2:

Hi Tony - 

Is there a use case (and if so can you share it please) where we need to have the same bundle accessing different revisions of the yang file? I am hesitant to keep multiple projects around for the same logical unit (i.e. a yang file) as it will add overhead and maintenance costs in general. Unless of course if we have a good reason. :)

When I looked through the yang bundles it looked like we were setting our selves up to publish versions of the bundles to a repo and then the projects can pull in what ever versions they want (even multiple versions). The only limitation in this design (might be - not 100% positive on this) is that the same bundle can not pull in two different versions of the same bundle. Do you have a use-case for this that you can share? If not I would propose that we only keep a single version of the yang files (bundles) in source, and let project pull down older versions from the repos as needed (as opposed to keeping an all encompassing master bundle, or multiple bundles - one per version.

For example, we currently have the following "published" version (at least I assume we are auto publishing artifacts?)

<artifactId>ietf-yang-types</artifactId>
 <version>2010.09.24.4-SNAPSHOT</version>

So by default the controller can pull in that version. Now if we update the yang types bundle to the newest version and the controller still needs the older version they would just keep the old version in their dependencies. If they needed both, they would need to ship both (need to investigate how bundles are packaged for shipping) but the bundle that needed the newer version would use the newer version, but everything else could use the older version etc. So in a running OSGi container you would see this:

76	ACTIVE      org.opendaylight.yangtools.model.ietf-yang-types_2010.9.24.4-SNAPSHOT
182	ACTIVE      org.opendaylight.yangtools.model.ietf-yang-types_2013.7.15.1-SNAPSHOT

Doing things this way would likely minimize the "cost" and confusion of maintaining the yang models, while still providing access to all of the old versions. And since all of the code is auto generated off of a static yang model (with a revision date) we know things are not changing, so there is no harm in just keeping the old bundle around on a public repo and not building it every time...

It may be worth a few of us jumping on a hangout/web-ex to discuss the particulars as I am not sure I am being 100% clear here.

Thank you!

Devin

Patch-set: 2