Update patch set 2
Patch Set 2:

Hi Robert, 

I am not sure I 100% follow your proposal. Are you suggesting that we need to keep unique bundles around for each version of the yang file?

Also, my approach would not force people to update to a new yang file, assuming we retain the older versions of the bundle on the maven repository. You are correct thought that we wont be able to provide code updates to the generated code for older yang models with this approach, without maintaining different streams for the different versions.

Patch-set: 2