Update patch set 2
Patch Set 2:

I will need to digest this quite a bit. Yes, we are invested in #1, so we do need a short-term solution for this, which I think should be a separate bundle.

We do need a long-term solution at least designed by the time we ship Helium.

We will breaking the binding part (e.g. Java APIs) into the MD-SAL project, so they will not clutter yangtools, which will expose option #3 (yang-data-api) and the schema metadata (yang-model-api). Inter-model compatibility will not be handled at this layer and will be pushed out towards individual areas of application, who are better-equiped to handle such things.

This will also mean the model bundles will need to be be broken up, disconnecting the models and the binding itself. 

I think $YangModuleInfo (which is a reflection of yang-model-api) will need to be retained, exposed and registered as an OSGi service to correctly interact with classpath, though.

Patch-set: 2
Label: Code-Review=0