Update patch set 1
Patch Set 1:

The hand coded-code is the code that implements the generated APIs. If the current yang model version has been replaced with a new version, the code that uses the APIs breaks. 

If you need a new rev of the model, maybe we can check it in side by side with the existing model, then you explicitly use the new revision

Patch-set: 1