Create patch set 3
Uploaded patch set 3.

Patch-set: 3
Subject: Fix for bug 838 and bug 839. When codecs serialize the Java objects to CompositeNode, then the parent's namespace is passed down to the children (i.e. Flow's namespace is passed down to all the children such as instructions). In order for this to be consistent, yangtools must make case augmentations adopt the namespace of its parent. https://git.opendaylight.org/gerrit/#/c/6666/ is the controller-side of this fix.  This is the yangtools-side.
Commit: 7b0b12a389181f4e97641b0482b4540cd7b0fb72