BUG-994: introduce ObjectCache for QNameModule/QName
This introduces caching for QNameModule using object-cache-api. This
will allow for efficient object reuse, at the cost of requiring a cache
implementation being available on the class path.

The use of cache has been not propagated to the static creator methods,
simply because the caching lookup introduces a cost which is not
acceptable in all contexts.

Change-Id: I5af9c7e54c726c05a5824521f5ed5183dc4899b4
Signed-off-by: Robert Varga <rovarga@cisco.com>