Bug 1027: Improved instance identifier codec for augmentations
InstanceIdentifierCodecImpl
Updated instance identifier codec to also create context
for child nodes of augmentation when augmentation is
for first time encountered in new context.

Improved structure of instance identifier codec to be bit
more readable and which allows for reuse.

LazyGeneratedCodecRegistry
Updated AugmentableDispatchCodec to try serialization
of InstanceIdentifier to augmentation when new
location is discovered and report back if serialization
failed.

Added documentation to adaptForPathImpl, renamed parameters
to be more readable.

Change-Id: Ie55b163a4ab617b82dd89ed85a504e104f147fe8
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>