BUG-994: introduce SchemaPath.create() and friends
This introduces a proper static constructor method, which does not
necesssarily have to allocate a new object. It also does not require an
explicit list as initializer, giving more flexibility to both users and
implementation itself.

Also introduces a couple of utility methods which make manipulating
SchemaPath easier.

This patch obsoletes the public contructor and getPath() method.

Change-Id: I4ec49d4336f4c30a9732c0bd95650e19c5d5e3f2
Signed-off-by: Robert Varga <rovarga@cisco.com>