Update patch set 1
Patch Set 1:

The OF 1.0 spec says "The vlan_vid field is 16 bits long, when an actual VLAN id is only 12 bits. The value 0xffff is used to indicate that no VLAN id was set."

Also in MatchV10Builder.setDlVlan(), we have the set the range 0-65535.

For  VTN distribution, to get the flow written, VlanId needs to be set to 0FFFF.

Patch-set: 1
Reviewer: Gerrit User 1676 <1676@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0