Update patch set 1
Patch Set 1:

To clarify: the value 0xFFFF is a value with specific meaning withing the OF1.0 encoding and as such needs to be parsed and semantically translated to proper model semantic -- which is the ID leaf not being present.

If we were to change the range here, that would really mean that 65535 is a valid VLAN ID -- which is not true.

Patch-set: 1