Update patch set 2
Patch Set 2: Code-Review-1

(4 comments)

Mostly minor comments / questions at this point. -1 because I believe we need to expand on the comments a little while we are in these classes (noted locations). Additionally, as a proof of concept to validate these changes I hope to remove the synchronization from the swagger UI and load those pages, and ensure that we stop seeing the visibility issues. I will try to get to that by the time we resubmit a clean up of comments.

Patch-set: 2
Label: Code-Review=-1