Update patch set 1
Patch Set 1:

I think the term "Documented" may be unclear to people. I was thinking of "Operation" - it's a more generalized term for "Rpc".

So we could have OperationError and OperationException. RpcError could derive from OperationError, to keep compatibility, or maybe we eliminate RpcError altogether. RpcResult, RestconfDocumentedException and NetconfDocumentedException could all be changed to use OperationError (and also the new TransactionCommitFailedException and ReadFailedException).

Patch-set: 1