Update patch set 4
Patch Set 4:

So far this looks good, but does this actually work? The reason I am worried is that the apply code assumes nodes are TreeNodes are present for present data, which is not a valid assumption with this patch. So if I write "bar/baz" into /foo, it is going to write only the index for /foo. Then if I try to write "baz" into /foo/bar -- will it understand that there is a pre-existing node for /foo/bar?

Patch-set: 4
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0