Update patch set 3
Patch Set 3:

It's hard to read the xtend code, but it appears you still have a synchronized block in there.  You don't want that in a static initializer.  The correct code would simply initialize the variable without any locks, and the semantics of static variable initialization in Java will take care of the rest.

Patch-set: 3
Label: Verified=0