Fix bug in BindingRuntimeContext#getSchemaDefinition
The PreConditions in BindingRuntimeContext#getSchemaDefinition need to check:

   Preconditions.checkArgument(!Augmentation.class.isAssignableFrom(cls));

instead of

   Preconditions.checkArgument(Augmentation.class.isAssignableFrom(cls))

Also made a minor change to YangToSourcesProcessor, used by the yang
maven plugin, to alleviate unit tests failures when running in Eclipse when a
test tries to load the SchemaContext for a DataObject class. It ends up
looking for the yang file under target-ide/classes/META-INF/yang but
it's under target. It's copied there by YangToSourcesProcessor but it
hard-codes the build output dir to "target". I kept this behavior but
also copied the yang files to the actual build output dir that gets
passed in (it's "target-ide" when in Eclipse) if it differs from
"target".

Change-Id: I79768850be372f020db47847b9cc45932d6ae2b1
Signed-off-by: tpantelis <tpanteli@brocade.com>