Change yang-maven-plugin to write yang files to build output
Made a minor change to YangToSourcesProcessor, used by the yang
maven plugin, to alleviate unit tests failures when running in Eclipse
when a test tries to load the SchemaContext for a DataObject class. It ends up
looking for the yang file under target-ide/classes/META-INF/yang but
it's under target. It's copied there by YangToSourcesProcessor but it
hard-codes the build output dir to "target". I kept this behavior but
also copied the yang files to the actual build output dir that gets
passed in (it's "target-ide" when in Eclipse) if it differs from
"target".

Change-Id: I84cb47e2c90f99a8dba70c8e92a9720d8617e58b
Signed-off-by: tpantelis <tpanteli@brocade.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>