Bug 1559: Fixed resolving of used extensions.
This is hotfix in Yang Parser Grammar. As it turnded out Yang parser didn't count with
use of extensions outside of body statements. Now yang parser is capable to parse yang files with used
extensions outside of body statements. The recommendation for future is to completely rewrite yang parser
grammar in general.

Change-Id: I65e61ef712fb499fe543753404a805a4dd14cf85
Signed-off-by: Lukas Sedlak <lsedlak@cisco.com>