Use xml.bind Base64 decoder instead of Guava
As it turned out Guava Base64 decoder is not
compatible with XML encoding, since it is much
stricter and throws error on new lines which
are pretty common in XMLs with Base64 encoded
data.

javax.xml.bind package provides same static
method for parsing string into bytes,
but with XSD semantics for Base64 encoding
which are compatible with XML way
of encoding data.

Change-Id: I5ef4c0b2ba521472b8caacdff1f44ff17d660eb3
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>