Bug 1514: Correctly handle ModuleOrModuleBuilder in ModuleDependencySort
ModuleDependencySort.processDependencies(...)

Turns out allNS.get(ns) can return objects of type ModuleOrModuleBuilder
(not just Module or ModuleBuilder types).

Now we check for that case, and unpack it to a Module or ModuleBuilder.
The rest of the logic then continues as normal.

Also added a log for getting a module with no name so as to catch
similar issues in the future.

Change-Id: I956c334c5dd10b51bec01d409f2433229d05f767
Signed-off-by: Ed Warnicke <eaw@cisco.com>