Update patch set 6
Patch Set 6:

FieldUpdater (since is part of JRE) does not use normal reflection, which is available via public java APIs, but uses JVM implementation private reflection (Unsafe), which pairs normal update of field so that cost will not be paid. Also Netty project uses FieldUpdaters instead of Atomics.
If memory is not problem, but performance we should not look even at AtomicLong or AtomicFieldUpdater, but java.util.concurrent.atomic.LongAdder (which was released also as public domain and has also JVM 7 port in opensource).

Patch-set: 6