Update patch set 5
Patch Set 5:

I created a test that uses nThreads to call addDuration a million times concurrently, waits for the threads to complete and captures the elapsed time. As expected the synchronized and concurrent versions perform similarly with 1 thread (uncontended) and the concurrent version outperforms as you increase the # of threads.

However, with running the test with multiple iterations (say 10 times), I usually see at least one outlier whose elapsed time is significantly higher, sometimes by a factor of 10. So, e.g., if the elapsed times for 9 runs is around 150 ms each, I'll see an outlier of 300 ms or 500 ms or even > 1 s. Now the test adds durations in increasing numbers from 0 to 1M so the largest duration is constantly changing which results in a new DurationWithTime instance each time. So eventually GC kicks in to clean them up and pauses the JVM. 

This is worst case. More commonly the longest/shortest durations will update less frequently. This could cause problems over time too if most DurationWithTime instances survive the young generation long enough to move to the old generation. In that case we may get a build up of instances in the old generation which may lead to significant GC pauses.

So it's a trade-off. The current code incurs 1 volatile write instead of 2 but incurs an object creation and resulting GC overhead. I tested w/o the object creation and making both duration and timeMillis volatile. While it's slightly slower, the elapsed times are relatively consistent. I prefer this approach b/c the behavior is relatively predictable and we avoid unpredictable GC interaction. Also having to update the 2nd timeMillis volatile will occur less frequently.

Patch-set: 5
Label: Verified=0