BUG-1431: make sure (Yang)InstanceIdentifier is Serializable
This works around the fact that an Iterable cannot be serialized by
ecplicitly iterating through it on the write path. On the read side, we
instantiate an ImmutableList and set that to be the path. This also
means that YangInstanceIdentifier can have its legacyPath set for free
-- so we do that.

Bumps the serialization format for InstanceIdentifier and
YangInstanceIdentifier to 2 and 3 respectively.

Change-Id: I668e2aeebde20b99cfe3b41594b3925697295d05
Signed-off-by: Robert Varga <rovarga@cisco.com>