BUG 1975 - yang unkeyed list is transformed to map node
Parsers and serializer for base and for concrete dom type were
added/modified to make it possible parse|serialize unkeyed list to|from
UnkeyedListNode.

!!REMARK!!
this change is tied also to project yang-data-operations which probaly
isn't used.
work on this issue will be continue 2 week leter (after 9. october 2014)
accroding to ttkacik

Change-Id: Ia3a9554dea00bb9d424495a2b6ed1f96765be738
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>