Introduce features-builder
The planned rework of version handling requires the features artifacts
to stop relying on maven properties for versions. This patch introduces
a base pom (e.g. suitable for being a parent), which provides the
versioning functionality based of dependency tree.

The user will typically use features-builder as their parent pom,
activate the plugins mentioned in pluginManagement and provide a
template features file, where the required version is marked by a
{{VERSION}} placeholder. The location of this template is hard-coded to
src/features, but the file name is controlled by features.file property
(features.xml) by default.

The build process' plugin wiring will take care of processing the
template and looking through the artifact's dependency tree. If a
required version is not found, the build is forced to fail.

Change-Id: If3252ffe8ca2b099ffd10c1cc37b47b3a1ac7698
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>