WIP: Schema Derived Data Tree
Yangtools and controller currently has several Tree structures
in which nodes are addressed by YangInstanceIdentifier and
are derived from SchemaContext programaticly only.

This patch proposes set of utility interfaces and classes
which whould ease implementation of such trees,
since they allow to cache model-derived data / algorithms
in tree structure which is similar to Normalized Node data
structures.

This allows for simultaneous walks in both trees, use
same algorithm for fetching data to fetch metadata and
so on.

Change-Id: Id5e41757e8a0042da28073dd851d1a222a1f3cf4
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>