Drop a FIXME for partially-overwritten writes
When a user performs a write, followed by a write which changes some
parts of the first one, we arrive at the resulting data in the prepare()
path. That is not right, as we can perform the merge in the seal() path.

The difference is that in MD-SAL the prepare() path is single-threaded,
while the seal() and read() path are executed in the application thread.
We really want to charge application-sourced mess to application
threads.

Change-Id: Iab2814f616ec8ca0d184513e5c8d3a3b8b1f2174
Signed-off-by: Robert Varga <rovarga@cisco.com>